finetune/requirements.txt
absl-py==1.2.0
tqdm==4.65.0
transformers==4.28.1
datasets==2.11.0
huggingface-hub==0.13.4
accelerate==0.18.0
